refactor(xod-client): extract some functions into new `actionUtils.js`, get rid of excessive comparation
